pyproject.toml
[project]
name = "bookscrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.3",
    "fastapi>=0.122.0",
    "fastapi-cli>=0.0.16",
    "lxml>=6.0.2",
    "pymongo[srv]>=4.15.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.34.0",
    "black>=25.11.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.14.6",
]
